test: S3Proxy 단위 테스트 작성
